% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Boik.test}
\alias{Boik.test}
\title{Boik's LBI test}
\usage{
Boik.test(x, nsim = 1000, ...)
}
\arguments{
\item{x}{A b-by-t data matrix, which rows corresponding
to b-block effects and columns are t-treatment effects.}

\item{nsim}{Number of simulation for compueting exact p-value. The defaut value is 1000.}
}
\value{
Exact and asymptotic p-value for input
}
\description{
Computes LBI's test statistics \eqn{(tr(R'R))^2/(p tr((R'R)^2))}, where R is residual
matrix of input data matrix under additivity assumption, and returns corresponung p-value.
}
\examples{
\dontrun{this is an example}
data(MVGH)
Boik.test(MVGH, nsim=10000)
}
\references{
Boik, R. J. (1993a). Testing additivity in two-way classifications
 with no replications: the locally best invariant test. Journal of Applied
 Statistics 20(1): 41-55.
}
\author{
Zahra. Shenavari, ....
}
