% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{combinep}
\alias{combinep}
\title{Combined several interaction tests}
\usage{
combinep(x, nsim = 500, nc0 = 10000, ...)
}
\arguments{
\item{x}{A data matrix in two-factor analysis}

\item{nsim}{Number of simulation for compueting exact p.value}

\item{dist}{If dist="sim", we used Monte Carlo simulation for estimating exact p-value,
 and if dist="asy", the p.values is
estimated from an asymptotic distribuion. The defaut value is "sim".}
}
\value{
A p.value for input
}
\description{
Reports p-values tests for non-additivity developed by Boik (1993a), Piepho (1994),
Kharrati-Kopaei and Sadooghi-Alvandi (2007), Franck et al. (2013), Malik et al. (2016)
and Kharrati-Kopaei and Miller (2016). In addition by use of four combination methods:
Bonferroni, Sidak, Jacobi expantion, and Gaussian Copula combined reported p-values.
}
\details{
If rows numer(b) of data matrix is less than it's columns number(t) we
 transpose data matrix. In addition requires that data matrix has more than two
  rows or columns.

Needs "mvtnorm" packages
}
\examples{
\dontrun{this is an example}
data(cnv6)
combinep(cnv6,nsim=500,nc0=10000)
}
\references{
Shenavari, Z.,Kharrati-Kopaei, M. (2018). A Method for Testing Additivity
in Unreplicated Two-Way Layouts Based on Combining Multiple Interaction Tests.International
Statistical Review
}
\author{
Zahra. Shenavari, ...
}
