% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Franck.test}
\alias{Franck.test}
\title{Franck (2013) et al. test for interaction}
\usage{
Franck.test(x, nsim = 1000, dist = "sim", ...)
}
\arguments{
\item{nsim}{a numeric value, the number of Monte Carlo samples for computing an exact Monte Carlo p-value. The default value is 10000.}

\item{dist}{a character, if dist="sim", a Monte Carlo simulation is used for calculating exact p-value,
and if dist="asy", the Bonferroni-adjusted p-value is calculated. The default is "sim".}

\item{\eqn{x}}{numeric matrix, \eqn{b \times a} data matrix where the number of rows and columns are corresponding to the block and treatment levels
, respectively.}
}
\value{
pvalue, calculated p-value.

nsim, the number of Monte Carlo samples that are used to estimate p-value.

statistic, the value of test statistic.
}
\description{
This function calculates Franck (2013) et al. test statistic,ACMIF, and corresponding p-value.
}
\details{
Franck et al. (2013) derived a test statistic based on the “hidden additivity” structure.
 They defined this structure as “the levels of one factor belong in two or more groups such that within each group the effects of the two factors are additive but the groups may interact with the ungrouped factor”.
 To detect hidden additivity, Franck et al. (2013) divided the table of data into two sub-tables and an interaction F-test was developed.
 Then, they performed a search over all possible configuRDWWns of data and used the maximum of the interaction F-RDWWs as a test statistic. The hypothesis of no interaction is rejected when the maximum interaction F-RDWW is large.
 Note that, if rows number, \eqn{b}, of data matrix is less than the columns number, \eqn{a}, 
 the data matrix is transposed. Note that the this test method is powerful when there is a hidden additivity structure in the data set.
}
\examples{
\dontrun{this is an example}
data(CNV)
Franck.test(CNV,nsim=1000,dist = "sim")
}
\references{
Franck, C., Nielsen, D., Osborne, J.A. (2013). A method for detecting hidden additivity in two-factor unreplicated experiments.
 Computational Statistics and Data Analysis 67:95-104. 
 
 Franck, C., Osborne, J.A. (2016).  Exploring Interaction Effects in Two-Factor Studies using the hidden Package in R.
 R Journal 8 (1):159-172.
 
 Shenavari, Z., Kharrati-Kopaei, M. (2018). A Method for Testing Additivity in
 Unreplicated Two-Way Layouts Based on Combining Multiple Interaction Tests. International Statistical Review
 86(3): 469-487.
}
\author{
Shenavari, Z.; Haghbin, H.; Kharrati-Kopaei, M.; Najibi, S.M.
}
