% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{hiddenf.test}
\alias{hiddenf.test}
\title{Franck et al.'s test for interaction}
\usage{
hiddenf.test(x, nsim = 1000, dist = "sim", ...)
}
\arguments{
\item{x}{A b-by-t data matrix, which rows corresponding
to b-block effects and columns are t-treatment effects}

\item{nsim}{Number of simulation for compueting exact p-value}

\item{dist}{If dist="sim", we used Monte Carlo simulation for estimating exact p-value,
and if dist="asy", Bonferroni-adjusted p-value computed. The defaut value is "sim"}
}
\value{
A p-value for input
}
\description{
computes Franck et al's. (2013) test statistic,ACMIF, and corresponding p-value.
}
\details{
If rows numer(b) of data matrix is less than it's columns number(t) at fiest
 transposing data matrix. In addition requires that data matrix has more than two
  rows or columns
}
\examples{
\dontrun{this is an example}
data(cnv6)
hiddenf.test(cnv6,nsim=1000,dist = "sim")
}
\references{
Franck, C., Nielsen, D., Osborne, J. A. (2013). A method for detecting
hidden additivity in two-factor unreplicated experiments. Computational Statistics
 and Data Analysis 67:95-104.
}
\author{
Zahra. Shenavari, ...
}
