% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{interactionplot}
\alias{interactionplot}
\title{Interaction Plot}
\usage{
interactionplot(x, ...)
}
\arguments{
\item{x}{numeric matrix, \eqn{a \times b} data matrix where the number of row and column is corresponding to the number of factor levels.}

\item{...}{plot parameters}
}
\value{
An interaction plot for input
}
\description{
Interaction Plot
}
\examples{
\dontrun{this is an example}
data(CNV)
interactionplot(CNV)
}
\author{
Shenavari, Z.; Haghbin, H.; Kharrati-Kopaei, M.; Najibi, S.M.
}
