% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{kh.f}
\alias{kh.f}
\title{internal function for hidden and KKSA methods}
\usage{
kh.f(x, bl, tr)
}
\arguments{
\item{x}{A data matrix}
}
\description{
internal function for hidden and KKSA methods
}
\keyword{internal}
