% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{KKM.test}
\alias{KKM.test}
\title{Kharrati-Kopaei and Miller (2016) test for interaction}
\usage{
KKM.test(x, nsim = 1000, nc0 = 10000)
}
\arguments{
\item{x}{numerix matrix, \eqn{b \times a} data matrix where the number of rows and columns are corresponding to the block and treatment levels, respectively.}

\item{nsim}{a numeric value, the number of Monte Carlo samples for computing an exact Monte Carlo p-value. The default value is 10000.}

\item{nc0}{a numeric value, the number of Monte Carlo samples for computing the unbiasing constant \eqn{c_0}. The default value is 10000.}
}
\value{
A list of consisting of:

pvalue, the calculated exact Monte Carlo p-value.

nsim, the number of Monte Carlo samples that are used to estimate p-value.

statistic, the value of test statistic.
}
\description{
This function calculates the test statistic for testing \eqn{H_0:} no interaction and corresponding Monte Carlo p-value
proposed by Kharrati-Kopaei and Miller(2016).
}
\details{
Kharrati-Kopaei and Miller(2016) proposed a test statistic for testing interaction
based on inspecting all pairwise interaction contrasts (PIC).
This test depends on an unbiasing constant \eqn{c_0} that is calculated by a Monte Carlo simulation.   
In addition, the null distribution of the test statistic is calculated by a Monte Carlo simulation. 
Note that this test procedure is powerful when significant interactions are caused by some data cells.
}
\examples{
\dontrun{this is an example}
data(RDWW)
KKM.test(RDWW,nsim=10000,nc0=10000)
}
\references{
Kharrati-Kopaei, M., Miller, A. (2016). A method for testing interaction in
 unreplicated two-way tables: using all pairwise interaction contrasts. Statistical
 Computation and Simulation 86(6):1203-1215.
  
 Shenavari, Z., Kharrati-Kopaei, M. (2018). A Method for Testing Additivity in
 Unreplicated Two-Way Layouts Based on Combining Multiple Interaction Tests. International Statistical Review
 86(3): 469-487.
}
\author{
Shenavari, Z.; Haghbin, H.; Kharrati-Kopaei, M.; Najibi, S.M.
}
