% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{KKSA.test}
\alias{KKSA.test}
\title{Kharrati-Kopaei and Sadooghi-Alvandi's test for interaction}
\usage{
KKSA.test(x, nsim = 1000, distr = "sim", ...)
}
\arguments{
\item{x}{A b-by-t data matrix, which rows corresponding
to b-block effects and columns are t-treatment effects}

\item{nsim}{Number of simulation for compueting exact p-value}

\item{dist}{If dist="sim", we used Monte Carlo simulation for estimating exact p-value,
and if dist="asy", Bonferroni-adjusted p-value computed. The defaut value is "sim"}
}
\value{
A p-value for input
}
\description{
computes Kharrati-Kopaei and Sadooghi-Alvandi's test statistics and corresponding p-value.
They proposed a test procedure for testing non-additivity based on splitting the b-by-t data
matrix into two non empty sub-matrixes by dividing the rows, each having at least two rows.
Suppose that b≥t and b≥4.Divided mean square errors for any two sub-matrixes such that
maximize its ratio and computed the corresponding p-value. Their test statistics are the
minimum value of p-values over all possible configurations: 2^(b-1)-b-1.
}
\details{
If rows numer(b) of data matrix is less than it's columns number(t) at fiest
 transposing data matrix. In addition requires that data matrix has more than three
  rows or columns
}
\examples{
\dontrun{this is an example}
data(impurity)
KKSA.test(impurity,nsim=1000,dist = "sim")
}
\references{
Kharrati-Kopaei, M., Sadooghi-Alvandi, S. M. (2007). A New Method for
 Testing Interaction in Unreplicated Two-Way Analysis of Variance. Communications
  in Statistics-Theory and Methods 36:2787–2803.
}
\author{
Zahra. Shenavari, ...
}
