% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{kpr}
\alias{kpr}
\title{This is an internal function which compute Kronecker product for PIC method}
\usage{
kpr(bl, tr)
}
\description{
This is an internal function which compute Kronecker product for PIC method
}
\keyword{internal}
