% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Malik_test.R
\name{Malik.test}
\alias{Malik.test}
\title{Malik (2016) et al. Test for Interaction}
\usage{
Malik.test(x, nsim = 10000, Elapsed.time = TRUE)
}
\arguments{
\item{x}{numeric matrix, \eqn{b \times a} data matrix where the number of rows and columns are corresponding to the block and treatment levels, respectively.}

\item{nsim}{a numeric value, the number of Monte Carlo samples for computing an exact Monte Carlo p-value. The default value is 10000.}

\item{Elapsed.time}{logical: if \code{TRUE} the progress will be printed in the console.}
}
\value{
An object of the class \code{ITtest}, which is a list inducing following components::
\item{pvalue}{The calculated exact Monte Carlo p-value.}
\item{nsim}{The number of Monte Carlo samples that are used to estimate p-value.}
\item{statistic}{The value of test statistic.}
\item{data.name}{The name of the input dataset.}
\item{test}{The name of the test.}
}
\description{
The Malik's (2016) et al. test statistics is calculated and the corresponding exact p-value is calculated by a Monte Carlo simulation.
}
\details{
Malik (2016) et al. proposed to partition
 the residuals into three clusters using a suitable clustering method like “k-means clustering”.
 The hypothesis of no interaction can be interpreted as the effect of the three
 clusters are equal. Therefore, the result of the test may depend on the method of clustering. In this package, clustering is done by 'kmeans' function in 'RcppArmadillo'. The 'speed_mode' parameter on the kmeans clustering was set as 'static_subset'.
 Note that the Malik's et al. test performs well when there are some outliers in the residuals; i.e. some cells produce large negative or positive residuals due to the significant interaction.
 Further, the distribution of the Malik's et al. test statistic is not known under additivity and the corresponding p-value is calculated by a Monte Carlo simulation.
}
\examples{
data(IDCP)
Malik.test(IDCP, nsim = 1000)

}
\references{
Malik, W.A., Mohring, J., Piepho, H.P. (2016). A
clustering-based test for non-additivity in an unreplicated two-way layout.
Communications in Statistics-Simulation and Computation 45(2):660-670.

Shenavari, Z., Kharrati-Kopaei, M. (2018). A Method for Testing Additivity in
Unreplicated Two-Way Layouts Based on Combining Multiple Interaction Tests. International Statistical Review
86(3): 469-487.
}
