% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{PIC.test}
\alias{PIC.test}
\title{Kharrati-Kopaei and Miller's test for interaction}
\usage{
PIC.test(x, nsim = 1000, nc0 = 10000, ...)
}
\arguments{
\item{x}{A b-by-t data matrix, which rows corresponding
to b-block effects and columns are t-treatment effects}

\item{nsim}{Number of simulation for compueting exact p-value. The defaut value is 1000}
}
\value{
An exact p-value for input
}
\description{
Computes the test statistics and corresponding p-value based on inspecting all
Pairwise Interaction Contrasts (PICs) for testing interaction
proposed by Kharrati-Kopaei and Miller(2016).
}
\examples{
\dontrun{this is an example}
data(ratio)
PIC.test(ratio,nsim=10000,nc0=10000)
}
\references{
Kharrati-Kopaei, M., Miller, A. (2016). A method for testing interaction in
 unreplicated two-way tables: using all pairwise interaction contrasts. Statistical
  Computation and Simulation 86(6):1203-1215.
}
\author{
Zahra. Shenavari, ...
}
