% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{piepho.test}
\alias{piepho.test}
\title{Piepho's third version test for interaction}
\usage{
piepho.test(x, nsim = 1000, ...)
}
\arguments{
\item{x}{A b-by-t data matrix, which rows corresponding
to b-block effects and columns are t-treatment effects}

\item{nsim}{Number of simulation for compueting exact p-value. The defaut value is 1000}
}
\value{
Exact and asymptotic p-value for input
}
\description{
Piepho (1994) proposed three test statistics.The third one is
based on Grubbs’ (1948) type estimator of variance for each level of block effect.
reports Piepho's test statistics and coresponding p-value by Monte Carlo datasets
or asympthotic distribution.
}
\examples{
\dontrun{this is an example}
data(MVGH)
piepho.test(MVGH,sim=1000)
}
\references{
Kharrati-Kopaei, M., Miller, A. (2016). A method for testing
interaction in unreplicated two-way tables: using all pairwise interaction contrasts.
 Statistical Computation and Simulation 86(6):1203-1215.
}
\author{
Zahra. Shenavari, ...
}
