% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDWW.R
\name{RDWW}
\alias{RDWW}
\title{Ratio of dry to wet wheat (RDWW).}
\format{
A matrix with four rows (Blocks) and four columns (Nitrogen Applied):
\describe{
  \item{Row1}{Block1}
  \item{Row2}{Block2}
  \item{Row3}{Block3}
  \item{Row4}{Block4}
  \item{Column1}{None}
  \item{Column2}{Early}
  \item{Column3}{Middle}
  \item{Column4}{Late}
}
}
\description{
This data set are about the ratio of dry to wet wheat of four different blocks and four times of nitrogen applied: None, Early, Middle, and Late.
The test of interaction between the blocks and the level of nitrogen applied is of interest.
}
\references{
\enumerate{
\item
  Ostle, B. (1963). Statistics in Research, Basic Concepts and Techniques for Research Works.
  2nd ed, p. 396. The Iowa State University Press.
}
}
