% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{CPI.test}
\alias{CPI.test}
\title{Combined p-value interaction test}
\usage{
CPI.test(x, nsim = 500, nc0 = 10000)
}
\arguments{
\item{x}{numeric matrix, \eqn{b \times a} data matrix where the number of rows and columns are corresponding to the block and treatment levels, respectively.}

\item{nsim}{a numeric value, the number of Monte Carlo samples for computing an exact Monte Carlo p-value. The default value is 10000.}

\item{nc0}{a numeric value, the number of Monte Carlo samples for computing the unbiasing constant \eqn{c_0}. The default value is 10000.}
}
\value{
A list of consisting of:

nsim, the number of Monte Carlo samples that are used to estimate p-value.

piepho.pvalue, the p-value of Piepho's (1994) test.

Boik.pvalue, the p-value of Boik's (1993) test.

Malik.pvalue, the p-value of Malik's (2016) et al. test.

KKM.pvalue, the p-value of Kharrati-Kopaei and Miller's (2016) test.

KKSA.pvalue, the p-value of Kharrati-Kopaei and Sadooghi-Alvandi's (2007) test.

Franck.pvalue, the p-value of Franck's (2013) et al. test.

Bonferroni, the combined p-value by using the Bonferroni method.

Sidak, the combined p-value by using the Sidak method.

jacobi, the combined p-value by using the Jacobi method.

GC, the combined p-value by using the Guassian copula.
}
\description{
This function reports the p-values of the tests for non-additivity developed by Boik (1993), Piepho (1994),
Kharrati-Kopaei and Sadooghi-Alvandi (2007), Franck et al. (2013), Malik et al. (2016)
and Kharrati-Kopaei and Miller (2016). In addition, it combines the p-values of these six methods into a single p-value as a test statistic for testing interaction.
There are four combination methods:
Bonferroni, Sidak, Jacobi expansion, and Gaussian Copula. The results of these four combinations are also reported.
}
\details{
If rows number ,\eqn{b} of data matrix is less than it's columns number, \eqn{a}, 
 the data matrix is transposed. In addition, this test procedure requires that the data matrix has more than two
 rows or columns. This function Needs "mvtnorm" package.
}
\examples{
\dontrun{this is an example}
data(RDWW)
CPI.test(RDWW,nsim=500,nc0=10000)

}
\references{
Shenavari, Z., Kharrati-Kopaei, M. (2018). A Method for Testing Additivity in
 Unreplicated Two-Way Layouts Based on Combining Multiple Interaction Tests. International Statistical Review
 86(3): 469-487.
}
\author{
Shenavari, Z.; Haghbin, H.; Kharrati-Kopaei, M.; Najibi, S.M.
}
