% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{interactionplot}
\alias{interactionplot}
\title{Interaction plot}
\usage{
interactionplot(x, ...)
}
\arguments{
\item{x}{numeric matrix, \eqn{b \times a} data matrix where the number of rows and columns are corresponding to the block and treatment levels, respectively.}

\item{...}{plot parameters}
}
\value{
An interaction plot for input
}
\description{
Interaction plot
}
\examples{
\dontrun{this is an example}
data(CNV)
interactionplot(CNV)
}
\author{
Shenavari, Z.; Haghbin, H.; Kharrati-Kopaei, M.; Najibi, S.M.
}
