% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{interactionplot}
\alias{interactionplot}
\title{Interaction plot}
\usage{
interactionplot(x, ...)
}
\arguments{
\item{x}{A data matrix in two-factor analysis}
}
\value{
An interaction plot for input
}
\description{
Interaction plot
}
\examples{
\dontrun{this is an example}
data(cnv6)
interactionplot(cnv6)
}
\author{
Zahra. Shenavari, ...
}
