% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Malik.test}
\alias{Malik.test}
\title{Malik (2016) et al. test for interaction}
\usage{
Malik.test(x, nsim = 10000)
}
\arguments{
\item{nsim}{a numeric value, the number of Monte Carlo samples for computing an exact Monte Carlo p-value. The default value is 10000.}

\item{\eqn{x}}{numeric matrix, \eqn{b \times a} data matrix where the number of rows and columns are corresponding to the block and treatment levels, respectively.}
}
\value{
A list of consisting of:

pvalue, the calculated exact Monte Carlo p-value.

nsim, the number of Monte Carlo samples that are used to estimate p-value.

statistic, the value of test statistic.
}
\description{
The Malik's (2016) et al. test statistics is calculated and the corresponding exact p-value is calculated by a Monte Carlo simulation.
}
\details{
Malik (2016) et al. proposed to partition
 the residuals into three clusters using a suitable clustering method like “k-means clustering”.
 The hypothesis of no interaction can be interpreted as the effect of the three
 clusters are equal.
 Note that the Malik's et al. test performs well when there are some outliers in the residuals; i.e. some cells produce large negative or positive residuals due to the significant interaction.
 Further, the distribution of the Malik's et al. test statistic is not known under additivity and the corresponding p-value is calculated by a Monte Carlo simulation.
}
\examples{
\dontrun{this is an example}
data(IDCP)
Malik.test(IDCP,nsim=10000)
}
\references{
Malik, W.A., Mohring, J., Piepho, H.P. (2016). A
clustering-based test for non-additivity in an unreplicated two-way layout.
Communications in Statistics-Simulation and Computation 45(2):660-670.

Shenavari, Z., Kharrati-Kopaei, M. (2018). A Method for Testing Additivity in
Unreplicated Two-Way Layouts Based on Combining Multiple Interaction Tests. International Statistical Review
86(3): 469-487.
}
\author{
Shenavari, Z.; Haghbin, H.; Kharrati-Kopaei, M.; Najibi, S.M.
}
