% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Malik.test}
\alias{Malik.test}
\title{Malik's test for interaction}
\usage{
Malik.test(x, nsim = 500)
}
\arguments{
\item{x}{A b-by-t data matrix, which rows corresponding
to b-block effects and columns are t-treatment effects}

\item{nsim}{Number of simulation for compueting exact p-value. The defaut value is 500}
}
\value{
An exact p-value for input
}
\description{
Repors the exact p-value from \code{\link{Malik et al. (2016)}}.They proposed to partition
the residuals(components of residual matrix R of input data matrix under additivity
assumption) into three clusters using a suitable clustering method like “k-means clustering”.
The hypothesis of no interaction can be interpreted as the effects of the three
clusters are equal.Compute Malik's test statistics and corresponding p-value by Monte Carlo simulation.
}
\examples{
\dontrun{this is an example}
data(impurity)
Malik.test(impurity,nsim=1000)
}
\references{
Malik, W. A., Mo ̈hring, J., Piepho, H. P. (2016). A
clustering-based test for non-additivity in an unreplicated two-way layout.
Communications in Statistics-Simulation and Computation 45(2):660-670.
}
\author{
Zahra. Shenavari, ..
}
